feat(snapshots): R2 VT layout goldens — @xterm/headless grid locks for the whole tree
Eight cases, twelve golden frames: banner at 80/40/160 (the 40-column
frame carries the D42 zero-row banner absence), a markdown turn, tool
cards collapsed+failed and ctrl+o expanded, the footer under full load at
80/40 (git badge and context yield under the S15 relief), the /model
panel's D30 dock takeover, the D48 adversarial corpus at 40 columns with
cell-level width verdicts (wide char, zero-width trailer, per-row width
sum), the D46 large-paste fold, and a pasted image turn.

- VtTerminal (tests/vt-terminal.ts): a FakeTerminal subclass mirroring
  every write into a real headless terminal; frames dump the whole buffer
  (scrollback included, cwd tokenized through every path-segment prefix —
  a truncation mid-segment still collapses to «cwd»/…). The alternate
  buffer is refused: Blue renders on the main screen, entering alt would
  itself be a regression.
- e2e-boot.ts: the boot infrastructure (bootBlue, helpers, module reset)
  extracted from e2e.spec.ts so the snapshot spec boots the same tree
  without importing a spec file; bootBlue grows a terminal injection
  option. The 114 e2e cases run unchanged.
- Golden frames are plain-text files under tests/golden/ via
  toMatchFileSnapshot — the PR diff is the reviewable layout statement.
  Regenerate with: pnpm vitest run packages/bundle/blue/tests/vt-snapshot.spec.ts -u
- Determinism: frames capture the VT buffer as-is after a two-phase
  quiesce (content, then 50ms-apart dumps agreeing); the git badge goes
  through the setGitCommandRunner seam; no timestamps, the title service
  is the e2e structural fake, spinner rows are idle-empty. A forced full
  repaint was tried and dropped: the whole-tree 40-column frame trips
  pi-tui's width guard mid-write under it (a real D48-family residue the
  component scans never see — recorded for its own fix).
- @xterm/headless 6.0.0 as a root devDependency (published 2025-12, no
  release-age or build-script concerns; version.spec only audits the
  dsh line, which is untouched).

Verified: 11/11 × 3 consecutive runs; negative tests (banner label,
footer gap) turn every affected frame red with readable diffs; the e2e
file and the full gate set stay green.
