+++ docs/blue-roadmap.md
@@ -158,7 +158,7 @@ alt-screen、主题切换、自定义键位、steer/cancel 的 UI、diff/termina
 | ✅ S36 | paste-image 失败分级提示 + 多类型可靠探测（**D49**）：六级失败分类 + `pasting image...`；列 MIME 后按 png>jpeg>webp>gif 读，空/坏/错 MIME representation 继续下一类型，store 最终交叉校验；`backend=auto|wayland|x11`（auto 原生优先、strict 禁跨协议）+ auto fallback 来源警示；3s SIGKILL timeout 按显示环境冷却 60s；Ubuntu Files 的 `text/uri-list`/GNOME copied-files 本地图片批次安全读取；真实 store e2e 逐字节回读（✅ 2026-08-22 真机验收通过——logout 复验后桌面复制路径确认，PR #21） | 0.5d | 无硬前置（interaction 独立面，worktree `worktree-s36-paste`） |
 | ✅ R1 | 钉版复核（✅ 2026-08-22 全量对齐 0.1.1-rc.2）：上游 rc.1→rc.2 对 Blue 消费面**零 API 变化**（9 包纯版本行、mcp-client 仅自家测试风格，实质变更在 image/attachment 管线内部与 deepseek llm 修复）；66 精确钉 + 35 caret peer + 181 excludes（pnpm 自动补录 4 条新包）+ HARNESS_LINE 同步，lockfile 最小演进（transitive 12 包经临时 overrides 升齐防双实例；jose 压回 6.2.9 避 minimumReleaseAge 24h 拒装——pnpm 11 默认供应链策略与 rc 线的交互记入 AGENTS）；**ci.yml 钉版单源化**至 `script/harness-line.mjs`（version.spec 增轻断言防字面量回潮）；session-title 上游 bug rc.2 仍未修、refresh 桥保留（两处 AGENTS 叙述更新）；全量回归 1844/1844 + smoke:happy 真机（rc.2 probe）全绿；验收项：真 API dogfood 双消息标题自纠（桥活体证据）随 PR 人工验收（已过，28e34a4）；**后续：harness-drift 每日监控落地**（同日 PR #25）：detect 脚本盯全系 next tag（等齐约束/跨 minor 只报不动）、BUMP_READY 由 dsh 无头 agent 改文件（version.spec 为完备证明）、workflow 确定性全套门禁绿才 push+PR upsert——R1 口径机械化，机制详根 AGENTS.md「Harness drift monitoring」 | 0.5d | G1 |
 >>>>>>> origin/master
-| R2 | 快照最小集：@xterm/headless VirtualTerminal，5-8 例核心帧（banner 首帧 / 对话+工具卡 / footer / 面板挂载 D30 形态 / CJK 宽度）；复审 D13 结论记 blue-decisions | 1-1.5d | G1（**S30 后拍**，键位面冻结） |
+| ✅ R2 | 快照最小集（✅ 2026-08-22）：`vt-snapshot.spec.ts` 八例十二帧（banner 80/40/160 三宽含 D42 零行形态 / markdown 一轮 / 工具卡折叠+失败同帧+ctrl+o 展开 / footer 全负载双宽含 S15 让宽形态 / ModelPanel D30 槽位替换 / CJK 40 列 D48 语料+逐 cell 宽度断言 / S32 大粘贴折叠 / 图片消息）；VtTerminal（FakeTerminal 子类双喂 @xterm/headless，bootBlue 增 terminal 注入缝并抽出 e2e-boot.ts——114 e2e 用例零语义变化）；黄金帧 `tests/golden/*.txt` 明文入库（toMatchFileSnapshot，`-u` 重生成，PR diff 即版式评审）；确定性三件套（quiescence 两段等待 / setGitCommandRunner 缝 / spec 顶层 chdir 恒定长度目录+cwd 逐段前缀 token——换目录自证过）；@xterm/headless 6.0.0 root devDep；**发现并记档**：强制全清重绘在整树 40 列帧触发 pi-tui 宽度守卫中途断写（D48 组合面残余，D13 复审记，快照缓冲直读规避待专项修）；D13 复审记 blue-decisions | 1-1.5d | G1 |
 | R3 | npm 发包：五包依赖序 core→interaction→transcript→app→bundle；dry-run 核 files/exports 子路径；dist-tag `rc` | 0.5d | G2 |
 | R4 | 安装路径验证：干净环境（临时 DSH_HOME）`dsh plugin add @dsh-blue/blue` → 启动冒烟；guide 补 registry 安装路径 | 0.5d | G3 |
 | R5 | dogfood 记录：registry 安装（非 dev link）跑一次完整真实任务归档；阻塞项回修重走 R3-R4 | 0.5d | G4 |
