+++ package.json
@@ -27,6 +27,7 @@
     "@deepseek-ai/dsh-llm-mock-server": "0.1.1-rc.2",
     "@types/node": "^22.20.0",
     "@vitest/coverage-v8": "^4.1.8",
+    "@xterm/headless": "6.0.0",
     "node-pty": "1.2.0-beta.15",
     "oxlint": "1.76.0",
     "tsdown": "^0.22.2",
