+++ packages/bundle/blue/tests/golden/banner-40.txt
@@ -1,0 +1,24 @@
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+ 
+╭──────────────────────────────────────╮
+│ >                                    │
+╰──────────────────────────────────────╯
+ mock  «cwd»
