+++ packages/bundle/blue/tests/golden/banner-80.txt
@@ -1,0 +1,24 @@
+ ╭────────────────────────────────────────────────────────────────────────────╮
+ │                                                                            │
+ │  ▐█▛█▛█▌  Welcome to Blue!                                                 │
+ │  ▐█████▌  Send /help for help information.                                 │
+ │                                                                            │
+ │  Directory: «cwd»                                      │
+ │  Model:     mock · mock                                                    │
+ │  Version:   0.1.0-rc.1                                                     │
+ │                                                                            │
+ ╰────────────────────────────────────────────────────────────────────────────╯
+
+
+
+
+
+
+
+
+
+ 
+╭──────────────────────────────────────────────────────────────────────────────╮
+│ >                                                                            │
+╰──────────────────────────────────────────────────────────────────────────────╯
+ mock  «cwd»
