+++ packages/bundle/blue/tests/golden/conversation-markdown-80.txt
@@ -1,0 +1,28 @@
+ ╭────────────────────────────────────────────────────────────────────────────╮
+ │                                                                            │
+ │  ▐█▛█▛█▌  Welcome to Blue!                                                 │
+ │  ▐█████▌  Send /help for help information.                                 │
+ │                                                                            │
+ │  Directory: «cwd»                                      │
+ │  Model:     mock · mock                                                    │
+ │  Version:   0.1.0-rc.1                                                     │
+ │                                                                            │
+ ╰────────────────────────────────────────────────────────────────────────────╯
+ 
+ ✨ show markdown
+ 
+ ● Title                                                                       
+                                                                               
+   • item                                                                      
+                                                                               
+   link (https://example.com)                                                  
+                                                                               
+   ```js                                                                       
+     const x = 1                                                               
+   ```                                                                         
+ 
+╭──────────────────────────────────────────────────────────────────────────────╮
+│ >                                                                            │
+╰──────────────────────────────────────────────────────────────────────────────╯
+ mock  «cwd»                                               
+                                                                         ctx 10
