+++ packages/bundle/blue/tests/golden/footer-full-40.txt
@@ -1,0 +1,24 @@
+ 
+ ✨ run
+ 
+ ● ok                                  
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+ 
+╭──────────────────────────────────────╮
+│ >                                    │
+╰──────────────────────────────────────╯
+ mock  «cwd»  ma...
+                   context: 1% (10/64k)
