+++ packages/bundle/blue/tests/golden/footer-full-80.txt
@@ -1,0 +1,24 @@
+ ╭────────────────────────────────────────────────────────────────────────────╮
+ │                                                                            │
+ │  ▐█▛█▛█▌  Welcome to Blue!                                                 │
+ │  ▐█████▌  Send /help for help information.                                 │
+ │                                                                            │
+ │  Directory: «cwd»                                      │
+ │  Model:     mock · mock                                                    │
+ │  Version:   0.1.0-rc.1                                                     │
+ │                                                                            │
+ ╰────────────────────────────────────────────────────────────────────────────╯
+ 
+ ✨ run
+ 
+ ● ok                                                                          
+
+
+
+
+ 
+╭──────────────────────────────────────────────────────────────────────────────╮
+│ >                                                                            │
+╰──────────────────────────────────────────────────────────────────────────────╯
+ mock  «cwd»  main [+2 -1 ↑1]                              
+                                                           context: 1% (10/64k)
