+++ packages/bundle/blue/tests/golden/paste-fold-80.txt
@@ -1,0 +1,24 @@
+ ╭────────────────────────────────────────────────────────────────────────────╮
+ │                                                                            │
+ │  ▐█▛█▛█▌  Welcome to Blue!                                                 │
+ │  ▐█████▌  Send /help for help information.                                 │
+ │                                                                            │
+ │  Directory: «cwd»                                      │
+ │  Model:     mock · mock                                                    │
+ │  Version:   0.1.0-rc.1                                                     │
+ │                                                                            │
+ ╰────────────────────────────────────────────────────────────────────────────╯
+ 
+ ✨ xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx
+    xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx
+    xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx
+    ... (13 more lines, 16 total, ctrl+o to expand)
+ 
+ ● got it                                                                      
+
+ 
+╭──────────────────────────────────────────────────────────────────────────────╮
+│ >                                                                            │
+╰──────────────────────────────────────────────────────────────────────────────╯
+ mock  «cwd»                                               
+                                                                         ctx 10
