+++ packages/bundle/blue/tests/golden/tool-card-expanded-80.txt
@@ -1,0 +1,40 @@
+ ╭────────────────────────────────────────────────────────────────────────────╮
+ │                                                                            │
+ │  ▐█▛█▛█▌  Welcome to Blue!                                                 │
+ │  ▐█████▌  Send /help for help information.                                 │
+ │                                                                            │
+ │  Directory: «cwd»                                      │
+ │  Model:     mock · mock                                                    │
+ │  Version:   0.1.0-rc.1                                                     │
+ │                                                                            │
+ ╰────────────────────────────────────────────────────────────────────────────╯
+ 
+ ✨ run both
+ 
+ 
+ ✓ Used long-output
+ 
+ 
+ ✗ Used boom · 1 line
+   Error: boom exploded
+ 
+ ● done                                                                        
+
+
+
+
+
+
+
+
+
+
+
+
+
+ 
+╭──────────────────────────────────────────────────────────────────────────────╮
+│ >                                                                            │
+╰──────────────────────────────────────────────────────────────────────────────╯
+ mock  «cwd»                                               
+                                                                         ctx 10
